Fix API inconsistencies with Chef 12.3.0
Chef 12.3.0 was released on April 28th 2015. In this commit is a small
change to the Gem package Resource that adds the attribute
`clear_sources`

The Gem Resource is coupled to the RubyGems provider which calls
clear_sources directly when installing packages.

The td agent cookbook provides their own TdAgentGem resource
(implemented as a Chef Lightweight Resource Provider). This resource
assigns itself a custom provider (The TdRubyGems provider, implemented
in libraries. This provider is a subclass of the main Chef Rubygems
provider and this is subject to API changes in the main Chef repo.

This commit adds the new clear_sources method into the custom Gem
Resource.

This is probably not an ideal solution, as it means tracking the main
API for the Chef Gem provider in this file and will break this package
everytime the main Chef repo is updated.

The better solution might be to find a way to subclass the main Chef Gem
package resource, so API changes are inherited.