added condition to service property
td-agent v4 does not ship /etc/init.d/td-agent file for distros that use
systemd (ubuntu focal, for example maybe also others). This makes
chef-client fail to run restart of service, since that was hardcoded to
use init.d script.

This adds a condistion to only set that property if the file actually
esists. If it doesn't exist, then chefs service-mechanism will default
to using whatever is available (systemd in this case) and the chef-run
succeeds to restart the service properly.